+++ app/src/main/java/com/example/foodapp/data/model/Menu.kt
@@ -1,13 +1,10 @@
 package com.example.foodapp.data.model
 
-import com.example.foodapp.utils.json_format.LocalTimeSerializer
-import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import java.time.LocalTime
 
 @Serializable
 data class Menu(
-    val id: Long?=null,
+    val id: Int?=null,
     val name: String="",
     val active: Boolean= true
 )
